+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This file is influenced by http://keepachangelog.com/.
 
 ### Fixed
 
+- Fixed leaking email addresses on password reset and resend email confirmation pages
+
 
 ## [v1.1.0] - 2016-08-23
 ### Added