+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is influenced by http://keepachangelog.com/.
 - Added two-factor authentication via SMS and authenticator to sign-in/up and verification
 - Verification of identity when accessing protected pages
 - Phone number verification when changing numbers
+- Simple feedback system and admin
 
 ### Changed